Sidebar, more Layer, GeoJSON
Added a new Sidebar. When the sidebar opens, the map moves. This solves
the Problem when you klick on a Point on the right side of the map, the
old sidebar overlays the point.

Some Layers are added to the Map.

The problem with missing Values and due to that the problem with
displaying the points is solved by another if request.